telephony_app/pyproject.toml
[tool.poetry]
name = "telephony_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
redis = "^4.5.4"
twilio = "^8.1.0"
vonage = "^3.5.1"
pyngrok = "^6.0.0"
python-dotenv = "^1.0.0"
vocode = "0.1.111a3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"